changed SessionNotFound handling to pick a new session from the pool
When a SessionNotFoundException is thrown, a new session is now picked
from the pool, instead of creating a new session on the server. This
reduces latency in calls that get an invalidated session from the pool.
It does however also require that the retry-logic for SessionNotFound is
put in a loop, as the second session returned from the pool could also
possibly be an invalidated session.